Fix "no null" error during replay when using GEF 3.16 or newer
The dispose listener of the graphical viewer is added to the
FigureCanvas via a method reference, rather than an anonymous class.

The reflective operation which tries to retrieve this viewer by looking
for the "this$0" variable no longer works, because this variable is now
called "arg$1".

With this change, we check for either of those variables.
Resolves #30