Fix "IOException: No preference value available"
Oomph logs a warning:
IOException: No preference value available for
preference:/instance/org.eclipse.oomph.workingsets/working.set.group
on application of any Preference context (even the one, where this value
is set).

It detects a momentary absence of a required value in a listener.
The fix  is not to clear values, that are present in a Preference
Context.