Do not wait for irrelevant jobs #95
UIJobCollector recognizes newly discovered jobs as required. As
ProblemViewSupport recreates its collector, it has to wait for any
preexisting jobs, even unrelated ones.

Other jobs are monitored by SWTUIPlayer, there is no need to wait for
them in ProblemViewSupport.

This should be handled in JobCollector extension, but for a quick
solution, we just fix ProblemViewSupport not to consider unrelated jobs.