Fix NPE #69
When an invalid *.target file is present in the workspace, PDE iterates
over it attempts to parse. This leads to exceptions in
getTargetDefinition() method and RCPTT handles that poorly when looking
up a needed target definition among the mix of valid and invalid ones.

The fix is to properly handle exceptions.

Fixes #69